+++ rosie/skills/use-unity-editor/resources/Serializer/SceneWriter.cs
@@ -109,6 +109,7 @@ namespace Rosie
                 ObjectEditor.ReadEdit(edit);
             }
             PrefabHydrator.EndEdits();
+            EditorSceneManager.SaveOpenScenes();
             shouldSerializeInNFrames = 2;
         }